Reorganize code into an src/ structure
Next supports src/ based project layouts without any reconfiguration.
Moving to this structure makes it slightly easier to keep all the code
together since the root of the repo is getting quite busy.
